requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==4.0.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dataclasses-json==0.5.7
entrypoints==0.4
faiss-cpu==1.12.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.33.4
idna==3.11
importlib_metadata==8.7.0
InstructorEmbedding==1.0.1
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.0.184
langchain-huggingface==1.0.0
langchain-text-splitters==1.0.0
langsmith==0.4.37
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
marshmallow-enum==1.5.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numexpr==2.14.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==0.27.6
openapi-schema-pydantic==1.2.4
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==3.20.3
pyarrow==21.0.0
pydantic==1.10.9
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
sentence-transformers==2.2.2
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.18.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.4.0
tokenizers==0.19.1
toml==0.10.2
toolz==1.0.0
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.29.0
triton==3.5.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
